Updated performance testing scripts
